experiments/mvp_adk_agents/mvp_adk_agent_with_ca/requirements.txt
google-adk>=1.3.0
google-auth
